Refactor Day 9 to use deque module
I wanted to extract the deque (double-ended queue) logic into a separate
library. While there is a deque library in the "pfds" package, it
doesn't support the 'rotate' operation like the Python deque structure
does, which was my inspiration, so I wrote it.

Rewriting the solution for Day 9 using the new deque library was a joy!