Day 14
For part 2, my solution was waaaaay slow. Just for a sanity check,
I also wrote the solution in Ruby, which ran in a little over 36
seconds. Then I came back to the Racket version and optimized it
so that it mutates a single string rather than building new ones
all the time. That helped, but the Racket version for part 2 still
takes over 15 minutes to run!

I don't know if I'm doing something stupid (extra object
allocations?), but I suspect that Ruby is just that much faster at
doing string manipulation than Racket.

    ~ time ruby day14part2.rb
    ruby day14part2.rb  36.55s user 0.04s system 100% cpu 36.588 total

    ~ time racket day14.rkt
    racket day14.rkt  971.86s user 3.85s system 100% cpu 16:15.47 total