Day 23
This involved a fair amount of manual work. I started out at a low
"resolution" looking for hot spots in the 3D space. Then I commented out
that code and copied the code to run just around the hot spot at a
slightly higher resolution.

I kept doing this until I was confident I found the right range and was
able to run it at the actual resolution of 1-to-1.

I wasn't even sure if this would work at first, so I was hesitant to
write a bunch of code with the assumption. But knowing what I know
now, this could be optimized to run fully in code.